move APIFnProperty to action_property.go
